pyproject.toml
[project]
name = "asset-information-extraction-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.5.2",
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
